env: auto extract when packaging
